index: split mutable index types to "mutable" module
Added pub(super) where needed or makes sense.
